Implemented confidence level for line artefact detector. #93
