+++ src/core/__init__.py
@@ -1,4 +1,4 @@
 from .water_detector import *
-from .line_artifact_detector import *
+from .line_artefact_detector import *
 from .artifact_detector import *
 from .color_diff import *
